# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lapras Sulfone Zinc API",
    "description": "Sulfone-Zinc: Registry API for CyanPrint",
    "contact": {
      "name": "kirinnee",
      "email": "kirinnee97@gmail.com"
    },
    "version": "1.0"
  },
  "paths": {
...
    "/api/v{version}/Template": {
      "get": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "Owner",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplatePrincipalResp"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplatePrincipalResp"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplatePrincipalResp"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/id/{userId}/{templateId}": {
      "get": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/slug/{username}/{name}": {
      "get": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateResp"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/id/{userId}": {
      "post": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplatePrincipalResp"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/slug/{username}/{templateName}/like/{likerId}/{like}": {
      "post": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "likerId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "like",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unit"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/slug/{username}/{templateName}/versions": {
      "get": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateName",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateVersionReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              }
            }
          }
        }
      }
    },
    "/api/v{version}/Template/id/{userId}/{templateId}/versions": {
      "get": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Limit",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Template"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "version",
            "in": "path",
            "description": "The requested API version",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "default": "1.0"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateVersionReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TemplatePrincipalResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "project": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "readme": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TemplateResp": {
        "type": "object",
        "properties": {
          "principal": {
            "$ref": "#/components/schemas/TemplatePrincipalResp"
          },
          "info": {
            "$ref": "#/components/schemas/TemplateInfoResp"
          },
          "user": {
            "$ref": "#/components/schemas/UserPrincipalResp"
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateVersionPrincipalResp"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateTemplateReq": {
        "type": "object",
        "properties": {
          "project": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "readme": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}